% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_split.R
\name{cluster_split}
\alias{cluster_split}
\title{Split clusters based on temperatures}
\usage{
cluster_split(pca.components, parameter.estimates, clusters, temperature,
  decay.step, max.clusters, num.cores)
}
\arguments{
\item{pca.components}{Expects a matrix with n rows as the cells and m columns as the principal components.}

\item{parameter.estimates}{Expects a list of parameters (mu, covariance, and weights) from initialize_gmm or maximization_step}

\item{clusters}{Expect the most likely cluster assignments}

\item{temperature}{Current temperature to use to split clusters}

\item{decay.step}{Data.frame with the first column as the step number and the second column as the temperature for that step}

\item{max.clusters}{Maximum number of clusters for the dataset}
}
\description{
Split clusters into new clusters based on temperature for the deterministic annealing algorithm
}
\keyword{internal}
