% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_function.R
\name{daGMM}
\alias{daGMM}
\title{Master wrapper for daGMM}
\usage{
daGMM(tsne.coordinates, pca.components, decay.step, max.iterations,
  delta.log.likelihood, max.clusters, tsne.plot.x.lims, tsne.plot.y.lims,
  num.cores)
}
\arguments{
\item{tsne.coordinates}{Expects a data.frame with two columns (tSNE_1 and tSNE_2, respectively) and rows of cells}

\item{pca.components}{Expects a matrix with n rows as the cells and m columns as the principal components.}

\item{decay.step}{Expects a data.frame of two columns, with the first as an index of temperature and the second as the temperatures for the determinisitic annealing}

\item{max.iterations}{Maximum number of iterations of the EM for each step}

\item{delta.log.likelihood}{Change in the log likelihood between steps of the EM for convergence}

\item{max.clusters}{Maximum number of clusters in the final step of the algorithm}
}
\description{
Takes input from the TSNE cluster coordinates, PCA components, decay step, maximum iterations, delta log likelihood for convergence at each step, and the maximum number of resulting clusters
}
