% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dragon-package.R
\docType{package}
\name{dragonsc-package}
\alias{dragonsc}
\alias{dragonsc-package}
\title{dragonsc: Single-cell RNAseq Deterministic Annealing Gaussian Mixture Modeling}
\description{
This package uses a deterministic algorithm inspired by "Deterministic Annealing for Clustering, Compression, Classification, Regression and Related Optimization Problems" in Proceedings of the IEEE, November 1998 by Kenneth Rose. It uses PCA components from single-cell expression and "temperatures" inspired by the physical process of annealing.
}
\author{
\strong{Maintainer}: Anthony Cillo \email{arc85@pitt.edu}

}
\keyword{internal}
