% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_function.R
\name{dragonsc}
\alias{dragonsc}
\title{DRAGONsc clustering algorithm}
\usage{
dragonsc(pca.components, temp.decay.steps, max.iterations,
  delta.log.likelihood, max.clusters, save.intermediate.files = FALSE,
  num.cores, verbose = FALSE)
}
\arguments{
\item{pca.components}{Expects a matrix with n rows as the cells and m columns as the principal components.}

\item{temp.decay.steps}{Temperature progressed used for the algorithm. Expects a two column data frame, with columns named "step" and "temperature"}

\item{max.iterations}{Maximum number of iterations of the expectation maximization at each temperature}

\item{delta.log.likelihood}{Change in the log likelihood between steps of the expectation maximization for convergence}

\item{max.clusters}{Maximum number of clusters in the final step of the algorithm}

\item{save.intermediate.files}{Whether to save files at each temperature of the algorithm. Default is FALSE, but can be very informative to view intermediate results.}

\item{num.cores}{Number of cores to use for the algorithm. Uses doParallel as a backend.}

\item{verbose}{Whether to display the current step of the algorithm and the log likelihood for each iteration of the expectation maximization algorithm. Default is FALSE.}
}
\description{
Takes input from, PCA components, temperature decay steps, maximum iterations, delta log likelihood for convergence at each step, maximum number of resulting clusters, and number of cores for parallelization. Has options for saving intermediate files and for verbose output.
}
