% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_step.R
\name{expect_max}
\alias{expect_max}
\title{EM step of the algorithm}
\usage{
expect_max(pca.components, clusters, parameter.estimates, max.iterations,
  delta.log.li, temp, num.cores, verbose)
}
\arguments{
\item{pca.components}{Expects a matrix with n rows as the cells and m columns as the principal components.}

\item{clusters}{Expect the most likely cluster assignments}

\item{parameter.estimates}{Expects a list of parameters (mu, covariance, and weights) from initialize_gmm or maximization_step}

\item{max.iterations}{The maximum number of iterations of the EM}

\item{delta.log.li}{Change in the log likelihood to satisfy convergence}
}
\description{
Wrapper to go between the expectation and maximization steps until the maximum iterations or the convergence of log likelihood
}
