% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_expectation_zero_temp.R
\name{expectation_step_zero}
\alias{expectation_step_zero}
\title{Deterministic annealing expectation}
\usage{
expectation_step_zero(pca.components, clusters, parameter.estimates,
  num.cores)
}
\arguments{
\item{pca.components}{Expects a matrix with n rows as the cells and m columns as the principal components.}

\item{clusters}{Expect the most likely cluster assignments}

\item{parameter.estimates}{Expects a list of parameters (mu, covariance, and weights) from initialize_gmm or maximization_step}

\item{num.cores}{Number of cores used for parallel calculations}
}
\description{
Uses a Euclidean distance metric to determine the distance of each cell from cluster centers, and uses a Gibbs sampler to evaluate the likelihood probability and posterior probability
}
\keyword{internal}
