% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_gmm.R
\name{initialize_gmm}
\alias{initialize_gmm}
\title{Initialization of the annealing model}
\usage{
initialize_gmm(pca.components, temperature, num.cores)
}
\arguments{
\item{pca.components}{Expects a matrix with n rows as the cells and m columns as the principal components.}

\item{temperature}{Initial temperature for annealing.}

\item{num.cores}{Number of cores used in parallel computations}
}
\description{
This function uses PCA components as input, and splits the data into two clusters at a given temperature. It also provides initial parameters for the expectation and maximization.
}
\keyword{internal}
