% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maximization_step.R
\name{maximization_step}
\alias{maximization_step}
\title{Maximization step of the algorithm}
\usage{
maximization_step(pca.components, clusters, parameter.estimates,
  posteriors, num.cores)
}
\arguments{
\item{pca.components}{Expects a matrix with n rows as the cells and m columns as the principal components.}

\item{clusters}{Expect the most likely cluster assignments.}

\item{parameter.estimates}{Expects a list of parameters (mu, covariance, and weights) from initialize_gmm or maximization_step.}

\item{posteriors}{Data.frame of priors from the expectation step.}

\item{num.cores}{Number of cores used for parallel calculations}
}
\description{
Use the expectation step to generate parameters of the mixture model
}
\keyword{internal}
